Add copy_on_swap attribute to ZoruaField
When a ZoruaField needs to have a packed repr for whatever reason, some
of its fields may not be referenceable by the auto implementation of
swap_bytes_mut.

In these cases, the the copy_on_swap attribute can be used as a way to
allow Copy fields to simply copy, swap, then set that field afterwards
in that implementation.